+++ Neuron.cpp
@@ -1,6 +1,6 @@
 #include "Neuron.h"
 
 Neuron::Neuron(int input_size) : input_size(input_size) {
-    weights = new int[input_size];
+    weights = new double[input_size];
     bias = 0;
 }
